Extract UI strings for localization #18
Refactor TimerPanel and TimerSequenceInspector components for improved localization and UI consistency
